+++ jfx/src/test/java/io/quarkiverse/mcp/servers/jfx/MCPServerJFXTest.java
@@ -1,0 +1,61 @@
+package io.quarkiverse.mcp.servers.jfx;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.net.URL;
+import java.time.Duration;
+import java.util.concurrent.CompletableFuture;
+
+import jakarta.inject.Inject;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import dev.langchain4j.agent.tool.ToolExecutionRequest;
+import dev.langchain4j.mcp.client.DefaultMcpClient;
+import dev.langchain4j.mcp.client.McpClient;
+import dev.langchain4j.mcp.client.transport.http.HttpMcpTransport;
+import io.quarkiverse.fx.FxApplication;
+import io.quarkus.test.common.http.TestHTTPResource;
+import io.quarkus.test.junit.QuarkusTest;
+import javafx.application.Application;
+
+/**
+ * Basic unit test suite for the {@link MCPServerJFX} class.
+ */
+@QuarkusTest
+public class MCPServerJFXTest {
+
+    @TestHTTPResource
+    URL url;
+
+    @Inject
+    ObjectMapper objectMapper;
+
+    private McpClient mcpClient;
+
+    @BeforeAll
+    static void startJFX() {
+        CompletableFuture.runAsync(() -> Application.launch(FxApplication.class));
+    }
+
+    @BeforeEach
+    void setUpMcpClient() {
+
+        mcpClient = new DefaultMcpClient.Builder()
+                .clientName("test-mcp-client-jfx")
+                .toolExecutionTimeout(Duration.ofSeconds(10))
+                .transport(new HttpMcpTransport.Builder().sseUrl(url.toString() + "mcp/sse").build())
+                .build();
+
+    }
+
+    //@Test disabling for now as getting npe on github actions
+    void launchCanvas() {
+        final var ret = mcpClient.executeTool(ToolExecutionRequest.builder().name("launchCanvas").arguments("{}").build());
+        assertThat(ret).startsWith("Canvas launched with dimensions:");
+    }
+
+}
