+++ shared/src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,1 +1,1 @@
-io.quarkus.mcp.servers.shared.McpCliConfigSource
+io.quarkiverse.mcp.servers.shared.McpCliConfigSource
